+++ Bootloader/Source/main.cpp
@@ -1,11 +1,16 @@
-#include <Uefi.h>
+#include <print.hpp>
+print Print;
+
+EFI_SYSTEM_TABLE* gSystem;
 
 extern "C"
 EFI_STATUS
 EFIAPI efi_main(EFI_HANDLE ImageHandle, EFI_SYSTEM_TABLE* System)
 {
-	System->ConOut->ClearScreen(System->ConOut);
-	System->ConOut->OutputString(System->ConOut, (CHAR16*)L"Hello, World from OxizeOS in UEFI!\r\n");
+	gSystem = System;
+	Print.SetStdout(System->ConOut);
+	Print.ClrScr();
+	printf("OxizeOS Bootloader V0.1.0002");
 
 	EFI_INPUT_KEY Key;
 	while (System->ConIn->ReadKeyStroke(System->ConIn, &Key) != EFI_SUCCESS);